intel/ibc: Use an int for num_bytes and num_comps in foreach callbacks
Using int8_t doesn't work because it's not actually large enough for
num_bytes.  Also, the smaller sizes likely aren't buying us anything
except MAYBE a tiny bit of stack space somewhere.  Meh.

+++ src/intel/compiler/ibc.h
@@ -516,7 +516,7 @@ typedef struct ibc_instr {
 } ibc_instr;
 
 typedef bool (*ibc_reg_ref_cb)(ibc_reg_ref *ref,
-                               int8_t num_bytes, int8_t num_comps,
+                               int num_bytes, int num_comps,
                                uint8_t simd_group, uint8_t simd_width,
                                void *state);
 bool ibc_instr_foreach_read(ibc_instr *instr, ibc_reg_ref_cb cb, void *state);
